refactor: ParticipatingOrganisations
